Add index checks

+++ db/client.py
@@ -23,12 +23,20 @@ class DBClient:
         self._client = meilisearch.Client(url=self._uri)
 
     def create_index(self, index_name: str) -> None:
+        if self._exists_index(index_name=index_name):
+            raise IndexAlreadyExistsError
         self._client.create_index(uid=index_name)
-        # TODO: すでにインデックスが存在する場合の警告
 
     def delete_index(self, index_name: str) -> None:
+        if not self._exists_index(index_name=index_name):
+            raise IndexNotExistsError
         self._client.delete_index(uid=index_name)
-        # TODO: インデックスが存在しない場合の警告
+
+    def _exists_index(self, index_name: str) -> bool:
+        indexes = self._client.get_indexes()["results"]
+        if index_name in [index.uid for index in indexes]:
+            return True
+        return False
 
     def add_document(self, index_name: str, document: Document) -> None:
         index = self._client.index(uid=index_name)
@@ -57,6 +65,18 @@ class DBClient:
 class TaskStatus(str, Enum):
     Succeeded = "succeeded"
     Failed = "failed"
+
+
+class IndexAlreadyExistsError(Exception):
+    def __init__(self) -> None:
+        super().__init__()
+        self.message = "Index already exists"
+
+
+class IndexNotExistsError(Exception):
+    def __init__(self) -> None:
+        super().__init__()
+        self.message = "Index not exists"
 
 
 class URLAlreadyExistsError(Exception):
